+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- geocode-service::findLocationAddress method
+- geocode-service::geocodeUrl method
+
 ## [0.8.3]
 ### Added
 - items-service::uploadResource added optional `filename` parameter